[BugFix] Support sub-1% sampling percent in TABLE SAMPLE and histogram analyze
On large tables the histogram analyze (ANALYZE FULL TABLE ... UPDATE HISTOGRAM)
computes a sample ratio below 1% from histogram_max_sample_row_count. The FE
truncated the percent to an integer, producing SAMPLE('percent'='0') which the
parser rejects with "invalid percent which should in (0, 100)".

Make the sampling probability a double end-to-end:
- thrift: add TTableSampleOptions.probability_percent_v2 (double); keep the
  legacy i64 probability_percent for backward compatibility with old BE.
- FE TableSampleClause: parse/store percent as double, allow fractional values
  in (0, 100), set both thrift fields, render percents without trailing ".0".
- FE HistogramStatisticsCollectJob: emit the true (possibly sub-1%) percent
  instead of truncating to an integer; fix the same latent truncation-to-0 in
  HyperStatisticSQLs by clamping to [1, 100] like its sibling builders.
- BE data_sample / segment_iterator / olap_chunk_source: read the new double
  field (falling back to the legacy field), use double throughout the bernoulli
  and histogram paths.

Add FE UTs, BE UTs and SQL tests covering sub-1% sampling.

+++ be/src/storage/rowset/data_sample.h
@@ -38,25 +38,27 @@ class DataSample {
 public:
     virtual ~DataSample() = default;
 
-    DataSample(int64_t probability_percent, int64_t random_seed)
+    DataSample(double probability_percent, int64_t random_seed)
             : _probability_percent(probability_percent), _random_seed(random_seed) {}
 
-    static std::unique_ptr<BlockDataSample> make_block_sample(int64_t probability_percent, int64_t random_seed,
+    static std::unique_ptr<BlockDataSample> make_block_sample(double probability_percent, int64_t random_seed,
                                                               size_t rows_per_block, size_t total_rows);
 
-    static std::unique_ptr<PageDataSample> make_page_sample(int64_t probability_percent, int64_t random_seed,
+    static std::unique_ptr<PageDataSample> make_page_sample(double probability_percent, int64_t random_seed,
                                                             size_t num_pages, PageIndexer page_indexer);
 
     virtual StatusOr<RowIdSparseRange> sample(OlapReaderStatistics* stats) = 0;
 
 protected:
-    int64_t _probability_percent;
+    // Sampling probability as a percent in the open range (0, 100). Stored as double so that sub-1%
+    // ratios (e.g. 0.5) on very large tables are preserved instead of being truncated to 0.
+    double _probability_percent;
     int64_t _random_seed;
 };
 
 class BlockDataSample final : public DataSample {
 public:
-    BlockDataSample(int64_t probability_percent, int64_t random_seed, size_t rows_per_block, size_t total_rows)
+    BlockDataSample(double probability_percent, int64_t random_seed, size_t rows_per_block, size_t total_rows)
             : DataSample(probability_percent, random_seed), _rows_per_block(rows_per_block), _total_rows(total_rows) {}
 
     StatusOr<RowIdSparseRange> sample(OlapReaderStatistics* stats) override;
@@ -95,7 +97,7 @@ struct SortableZoneMap {
 
 class PageDataSample final : public DataSample {
 public:
-    PageDataSample(int64_t probability_percent, int64_t random_seed, size_t num_pages, PageIndexer page_indexer)
+    PageDataSample(double probability_percent, int64_t random_seed, size_t num_pages, PageIndexer page_indexer)
             : DataSample(probability_percent, random_seed),
               _num_pages(num_pages),
               _page_indexer(std::move(page_indexer)) {}
